fix: relocate InsertOrUpdateOptions file with query-builder
